# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Smart Recruitment System API
  version: 1.0.0
  description: |
    REST API لمنظومة التوظيف الذكية. المصادقة عبر Bearer JWT.
    الواجهة الحالية تحاكي كل هذه المسارات محليًا (Pinia + localStorage)؛
    التبديل للحقيقي عبر VITE_USE_REAL_API=true دون تغيير واجهات المخازن.
servers:
  - url: http://localhost:8000/api/v1
    description: بيئة التطوير المحلية (Laravel)
security:
  - bearerAuth: []
paths:
...
  /direct-messages:
    get:
      tags: [Messaging]
      summary: كل رسائلي (مُرسَلة ووارِدة) لإعادة بناء المحادثات
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DirectMessage' }
    post:
      tags: [Messaging]
      summary: إرسال رسالة مباشرة لمستخدم (يصله لحظيًا عبر Socket.IO)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [recipientId, recipientName, body]
              properties:
                recipientId: { type: string }
                recipientName: { type: string }
                body: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DirectMessage' }
  /direct-messages/read:
    post:
      tags: [Messaging]
      summary: تعليم رسائل خيط (من طرف معيّن) مقروءة
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [peerId]
              properties:
                peerId: { type: string }
      responses:
        '204': { description: تم }
  /direct-messages/resolve/{slug}:
    get:
      tags: [Messaging]
      summary: يحلّ مالك صفحة تعريفية من slug (uuid واسم) لتوجيه «تواصل معي»
      parameters: [{ $ref: '#/components/parameters/slug' }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: [object, 'null']
                properties:
                  ownerId: { type: string }
                  name: { type: string }

  # ===== المستندات الخاصة العامة (blob) — تقابل Supabase account_states =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    DirectMessage:
      type: object
      properties:
        id: { type: integer }
        senderId: { type: string }
        recipientId: { type: string }
        senderName: { type: string }
        recipientName: { type: string }
        body: { type: string }
        created_at: { type: string, format: date-time }
        read_at: { type: [string, 'null'], format: date-time }
...